+++ express/src/decorators/middleware.ts
@@ -1,21 +1,23 @@
-import { getExpressMeta } from '../meta';
+import { getMeta, getMiddleware } from '../meta';
 
+/**
+ * Registers route middleware
+ * @deprecated
+ * @param {Function|Function[]} [middleware]
+ * @returns {(target:Function)=>void}
+ * @constructor
+ */
 export let Middleware = (middleware: Function|Function[]): MethodDecorator => {
   return (target: ExpressClass, propertyKey: string | symbol, descriptor: any) => {
-    let meta = getExpressMeta(target);
-    let _middleware: Function[];
+    console.warn('Deprecation: use route based middleware, @Middleware will be removed in 2.0.0, see doc -', target.constructor.name, propertyKey);
+
+    let meta = getMeta(target);
 
     if (!meta.middleware[propertyKey]) {
       meta.middleware[propertyKey] = [];
     }
 
-    if (typeof middleware === 'function') {
-      _middleware = [middleware];
-    } else if (Array.isArray(middleware)){
-      _middleware = (<Function[]>middleware).filter(md => typeof md === 'function');
-    }
-
-    meta.middleware[propertyKey].push(..._middleware);
+    meta.middleware[propertyKey].push(...getMiddleware(middleware));
 
     return descriptor;
   }